backend/pyproject.toml
[project]
name = "security-platform-backend"
version = "2.0.0"
description = "Backend da Security Platform com execução de ferramentas via Kubernetes"
authors = [{name = "Security Platform Team"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0",
    "sse-starlette==1.6.5",
    "kubernetes>=28.0.0",
    "psycopg2-binary>=2.9.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "models", "services"]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

